+++ stream-video-android-core/src/main/kotlin/io/getstream/video/android/core/call/connection/Subscriber.kt
@@ -1,23 +1,28 @@
 package io.getstream.video.android.core.call.connection
 
-import io.getstream.android.video.generated.models.VideoEvent
 import io.getstream.log.taggedLogger
 import io.getstream.result.flatMap
+import io.getstream.video.android.core.ParticipantState
 import io.getstream.video.android.core.api.SignalServerService
 import io.getstream.video.android.core.call.TrackDimensions
 import io.getstream.video.android.core.call.connection.utils.wrapAPICall
-import io.getstream.video.android.core.call.utils.stringify
-import io.getstream.video.android.core.events.ICETrickleEvent
-import io.getstream.video.android.core.model.AudioTrack
+import io.getstream.video.android.core.call.utils.TrackOverridesHandler
+import io.getstream.video.android.core.dispatchers.DispatcherProvider
 import io.getstream.video.android.core.model.IceCandidate
 import io.getstream.video.android.core.model.MediaTrack
 import io.getstream.video.android.core.model.StreamPeerType
-import io.getstream.video.android.core.model.VideoTrack
 import io.getstream.video.android.core.trySetEnabled
 import io.getstream.video.android.core.utils.safeCall
 import io.getstream.video.android.core.utils.safeCallWithResult
+import io.getstream.video.android.core.utils.safeSuspendingCallWithResult
 import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.retryWhen
+import kotlinx.coroutines.launch
 import org.webrtc.MediaConstraints
 import org.webrtc.MediaStream
 import org.webrtc.MediaStreamTrack
@@ -25,9 +30,14 @@ import org.webrtc.RtpTransceiver
 import org.webrtc.SessionDescription
 import stream.video.sfu.models.PeerType
 import stream.video.sfu.models.TrackType
+import stream.video.sfu.models.VideoDimension
 import stream.video.sfu.signal.ICERestartRequest
 import stream.video.sfu.signal.SendAnswerRequest
+import stream.video.sfu.signal.TrackSubscriptionDetails
+import stream.video.sfu.signal.UpdateSubscriptionsRequest
+import stream.video.sfu.signal.UpdateSubscriptionsResponse
 import java.util.concurrent.ConcurrentHashMap
+import io.getstream.result.Result
 
 internal class Subscriber(
     private val sessionId: String,
@@ -49,11 +59,14 @@ internal class Subscriber(
     maxBitRate = 0 // Set as needed
 ) {
     private val subscriberLogger by taggedLogger("Video:Subscriber")
-
+    companion object {
+        val defaultVideoDimension = VideoDimension(720, 1080)
+    }
     internal var onAddStream: ((MediaStream) -> Unit) = { _ -> }
 
     // Track dimensions state for this subscriber
     val trackDimensions = MutableStateFlow<Map<String, Map<TrackType, TrackDimensions>>>(emptyMap())
+    val subscriptions = MutableStateFlow<List<TrackSubscriptionDetails>>(emptyList())
 
     // Tracks for all participants (sessionId -> (TrackType -> MediaTrack))
     private val _tracks: ConcurrentHashMap<String, ConcurrentHashMap<TrackType, MediaTrack>> = ConcurrentHashMap()
@@ -158,6 +171,28 @@ internal class Subscriber(
         sfuClient.iceRestart(request)
     }
 
+    /**
+     * Sets the video subscriptions for the subscriber.
+     *
+     * @param useDefaults Whether to use the default tracks or not.
+     */
+    suspend fun setVideoSubscriptions(trackOverridesHandler: TrackOverridesHandler, participants: List<ParticipantState>, remoteParticipants: List<ParticipantState>, useDefaults: Boolean = false) : Result<UpdateSubscriptionsResponse> = safeSuspendingCallWithResult {
+        logger.d { "[setVideoSubscriptions] #sfu; #track; useDefaults: $useDefaults" }
+        var tracks = if (useDefaults) {
+            // default is to subscribe to the top 5 sorted participants
+            defaultTracks(participants)
+        } else {
+            // if we're not using the default, sub to visible tracks
+            visibleTracks(remoteParticipants)
+        }.let(trackOverridesHandler::applyOverrides)
+        subscriptions.value = tracks
+        val request = UpdateSubscriptionsRequest(
+            session_id = sessionId,
+            tracks = subscriptions.value,
+        )
+        sfuClient.updateSubscriptions(request)
+    }
+
     internal fun addTransceivers() {
         connection.addTransceiver(
             MediaStreamTrack.MediaType.MEDIA_TYPE_VIDEO,
@@ -167,6 +202,53 @@ internal class Subscriber(
             MediaStreamTrack.MediaType.MEDIA_TYPE_AUDIO,
             RtpTransceiver.RtpTransceiverInit(RtpTransceiver.RtpTransceiverDirection.RECV_ONLY)
         )
+    }
+
+    private fun defaultTracks(participants: List<ParticipantState>): List<TrackSubscriptionDetails> {
+        val otherParticipants = participants.filter { it.sessionId != sessionId }.take(5)
+        val tracks = mutableListOf<TrackSubscriptionDetails>()
+        otherParticipants.forEach { participant ->
+            if (participant.videoEnabled.value) {
+                val track = TrackSubscriptionDetails(
+                    user_id = participant.userId.value,
+                    track_type = TrackType.TRACK_TYPE_VIDEO,
+                    dimension = defaultVideoDimension,
+                    session_id = participant.sessionId,
+                )
+                tracks.add(track)
+            }
+            if (participant.screenSharingEnabled.value) {
+                val track = TrackSubscriptionDetails(
+                    user_id = participant.userId.value,
+                    track_type = TrackType.TRACK_TYPE_SCREEN_SHARE,
+                    dimension = defaultVideoDimension,
+                    session_id = participant.sessionId,
+                )
+                tracks.add(track)
+            }
+        }
+
+        return tracks
+    }
+
+    private fun visibleTracks(remoteParticipants: List<ParticipantState>): List<TrackSubscriptionDetails> {
+        val trackDisplayResolution = trackDimensions.value
+        val tracks = remoteParticipants.map { participant ->
+            val trackDisplay = trackDisplayResolution[participant.sessionId] ?: emptyMap()
+
+            trackDisplay.entries.filter { it.value.visible }.map { display ->
+                subscriberLogger.i {
+                    "[visibleTracks] $sessionId subscribing ${participant.sessionId} to : ${display.key}"
+                }
+                TrackSubscriptionDetails(
+                    user_id = participant.userId.value,
+                    track_type = display.key,
+                    dimension = display.value.dimensions,
+                    session_id = participant.sessionId,
+                )
+            }
+        }.flatten()
+        return tracks
     }
 
     override fun onAddStream(stream: MediaStream?) {
